fix(renderer): guard external link confirmation
